# openapi/openapi-payment.json
{
	"openapi": "3.1.0",
	"info": {
	  "title": "IWM Backend API",
	  "description": "Documentation for the modular monolith hexagonal backend APIs.",
	  "version": "0.1.0-SNAPSHOT"
	},
	"servers": [
	  {
		"url": "https://kernel-core.yowyob.com/kernel-api",
		"description": "Generated server url"
	  }
	],
	"security": [
	  {
		"ClientId": [],
		"ApiKey": []
	  },
	  {
		"ClientId": [],
		"ApiKey": [],
		"BearerAuth": []
	  }
	],
	"paths": {
...
	  "/api/payments/wallets/{walletId}/can-operate": {
		"get": {
		  "tags": [
			"payment-controller"
		  ],
		  "operationId": "canOperate",
		  "parameters": [
			{
			  "name": "walletId",
			  "in": "path",
			  "required": true,
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			},
			{
			  "name": "amount",
			  "in": "query",
			  "required": true,
			  "schema": {
				"type": "number"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "OK",
			  "content": {
				"*/*": {
				  "schema": {
					"$ref": "#/components/schemas/ApiResponseBoolean"
				  }
				}
			  }
			}
		  }
		}
	  },
...
	},
	"components": {
	  "schemas": {
...
		"ApiResponseBoolean": {
		  "type": "object",
		  "properties": {
			"success": {
			  "type": "boolean"
			},
			"data": {
			  "type": "boolean"
			},
			"message": {
			  "type": "string"
			},
			"errorCode": {
			  "type": "string"
			},
			"timestamp": {
			  "type": "string",
			  "format": "date-time"
			}
		  }
		},
...
	  },
	  "securitySchemes": {
		"ClientId": {
		  "type": "apiKey",
		  "description": "Client application identifier for server-to-server access",
		  "name": "X-Client-Id",
		  "in": "header"
		},
		"ApiKey": {
		  "type": "apiKey",
		  "description": "Client application secret for server-to-server access",
		  "name": "X-Api-Key",
		  "in": "header"
		},
		"BearerAuth": {
		  "type": "http",
		  "description": "JWT access token returned by /api/auth/login",
		  "scheme": "bearer",
		  "bearerFormat": "JWT"
		}
	  }
	}
  }